removed old merge in favor of c function
