+++ tests/test_itertools_recipes.py
@@ -50,7 +50,6 @@ def test_empty_input():
     # no need to test iter_except here
     # no need to test iter_subclasses here
     assert not iteration_utilities.last_true(empty)
-    assert list(iteration_utilities.merge(empty)) == []
     assert list(iteration_utilities.ncycles(empty, 10)) == []
     assert iteration_utilities.nth(empty, 10) is None
     assert list(iteration_utilities.take(
